functions/requirements.txt
firebase_functions~=0.1.0
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
markitdown[all]>=0.1.0
pdfminer.six>=20221105
pypdf>=3.0.0